+++ dask/dataframe/dispatch.py
@@ -15,6 +15,7 @@ categorical_dtype_dispatch = Dispatch("CategoricalDtype")
 concat_dispatch = Dispatch("concat")
 tolist_dispatch = Dispatch("tolist")
 is_categorical_dtype_dispatch = Dispatch("is_categorical_dtype")
+union_categoricals_dispatch = Dispatch("union_categoricals")
 
 
 def concat(
@@ -76,3 +77,8 @@ def categorical_dtype(meta, categories=None, ordered=False):
 def tolist(obj):
     func = tolist_dispatch.dispatch(type(obj))
     return func(obj)
+
+
+def union_categoricals(to_union, sort_categories=False, ignore_order=False):
+    func = union_categoricals_dispatch.dispatch(type(to_union[0]))
+    return func(to_union, sort_categories=sort_categories, ignore_order=ignore_order)
